refactor: 깨진 테스트 코드 수정
