+++ backend/src/main/java/edonymyeon/backend/service/ImageFileUploader.java
@@ -33,4 +33,8 @@ public class ImageFileUploader {
             throw new RuntimeException();
         }
     }
+
+//    public void setImageFileNameStrategy(final ImageFileNameStrategy imageFileNameStrategy) {
+//        this.imageFileNameStrategy = imageFileNameStrategy;
+//    }
 }
